requirements.txt

streamlit
pyairtable
python-dotenv
